Optimize MessagePackReader for contiguous memory
ReadOnlySequence<T> is expensive compared to Memory<T>, and this shows up markedly for small deserialization buffers (e.g. deserializing an integer).
This commit adds support for skipping ReadOnlySequence<T> entirely for cases where we have contiguous memory. We may lazily create the sequence if the caller demands it however.