Remove mutable fields/properties from TypelessFormatter
This change makes MessagePackSerializerOptions support subclassing while maintaining its immutable nature.

Also move the blacklist in TypelessFormatter to MessagePackSerializerOptions where it is now extensible/replaceable with a whitelist.
Also fix the short-name switch to store the cache in two different fields based on the switch to avoid cache corruption when the switch is changed after serializing some types.

Fixes #444
Fixes #567