Defend against stack overflow from deeply nested object graphs in msgpack sequence
Msgpack can very cheaply express a very deep object graph. For example a nested array takes only 2 bytes to express. This allows a short msgpack sequence to cause deserialization to crash from running out of stack space in a DoS attack.
The defense is to throw a catchable exception before the stack is blown by counting how deep the object graph is and proactively throwing when the graph depth exceeds some value.

This change adds functionality to the MessagePackSecurity class to help track and enforce this limit, and updates all the formatters that call other formatters (in a non-thunking way) to participate in this depth tracking.